+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add top articles and suggestions to welcome screen
 * Make ordering of suggestions match feedback site default (Hot Ideas, New Ideas, or Top Ideas)
 * Add instant answers to contact form
+* Fix a bug where it appeared you could vote on declined ideas
 
 ## 1.1.1 ##