+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## master ##
 
+* Track instant answers for analytics
+
+## 1.2.0 ##
+
 * Add top articles and suggestions to welcome screen
 * Make ordering of suggestions match feedback site default (Hot Ideas, New Ideas, or Top Ideas)
 * Add instant answers to contact form
 * Fix a bug where it appeared you could vote on declined ideas
 * Add support for helpdesk-only accounts
-* Track instant answers
 
 ## 1.1.1 ##